modify .y to better reveal FuncParams; tailor the tree; VS format before coding llvm; working on IR

+++ code/AST.hpp
@@ -18,10 +18,10 @@ public:
 
 /// NumberExprAST - Expression class for numeric literals like "1.0".
 class NumberExprAST : public ExprAST {
-  double Val;
+  int Val;
 
 public:
-  NumberExprAST(double Val) : Val(Val) {}
+  NumberExprAST(int Val) : Val(Val) {}
 
   Value *codegen() override;
 };
